GROW-45 rename and refactor Validator component
