--- src/ui/Validator/Validator.styles.ts
@@ -1,18 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  textInput: {
-    backgroundColor: 'white',
-    borderBottomWidth: 3,
-    borderColor: 'blue',
-    borderRadius: 5,
-    height: 40,
-    margin: 5,
-    padding: 15,
-    width: 200,
-  },
-  error: {
-    borderBottomWidth: 3,
-    borderColor: 'red',
-  },
-});
